Conflicts:
	src/applications/flexray/sink/FiCo4OMNeT_FRTrafficSinkAppBase.h
	src/applications/flexray/source/FiCo4OMNeT_FRTrafficSourceAppBase.h
	src/linklayer/can/FiCo4OMNeT_CanPortInput.h
	src/linklayer/can/FiCo4OMNeT_CanPortOutput.h
	src/linklayer/flexray/FiCo4OMNeT_FRPortInput.h
	src/linklayer/flexray/FiCo4OMNeT_FRPortOutput.h
	src/scheduler/can/FiCo4OMNeT_CanClock.h
	tests/fingerprint/fingerprints
	tests/smoke/smoketest

+++ src/scheduler/flexray/FiCo4OMNeT_SchedulerEvent.h
@@ -13,7 +13,7 @@ class SchedulerEvent : public SchedulerEvent_Base
     private:
         cGate *destinationGate_var;
     public:
-        SchedulerEvent(const char *name = NULL, int kind = 0) :
+        SchedulerEvent(const char *name = NULL, short kind = 0) :
             SchedulerEvent_Base(name, kind)
         {
             destinationGate_var = NULL;
